+++ client/src/pages/login.tsx
@@ -1,55 +1,9 @@
-import { Formik, Field, Form } from 'formik';
-import { FormControl, FormLabel, Input, FormErrorMessage, Button } from "@chakra-ui/react";
+import Login from "../components/Login/Login";
 
 export default function loginPage() {
-  function validateName(value) {
-    let error
-    if (!value) {
-      error = 'Name is required'
-    }
-    return error
-  }
-
   return (
-    <Formik
-      initialValues={{
-        email: '',
-        password: '',
-      }}
-      onSubmit={(values, actions) => {
-        console.log('values', values);
-      }}
-    >
-      {(props) => (
-        <Form>
-          <Field name='email' validate={validateName}>
-            {({ field, form }) => (
-              <FormControl isInvalid={form.errors.email && form.touched.email}>
-                <FormLabel htmlFor='email'>Email</FormLabel>
-                <Input {...field} id='email' placeholder='email' />
-                <FormErrorMessage>{form.errors.email}</FormErrorMessage>
-              </FormControl>
-            )}
-          </Field>
-          <Field name='password' >
-            {({ field, form }) => (
-              <FormControl isInvalid={form.errors.password && form.touched.password}>
-                <FormLabel htmlFor='password'>Password</FormLabel>
-                <Input {...field} id='password' placeholder='password' type='password' />
-                <FormErrorMessage>{form.errors.password}</FormErrorMessage>
-              </FormControl>
-            )}
-          </Field>
-          <Button
-            mt={4}
-            colorScheme='teal'
-            isLoading={props.isSubmitting}
-            type='submit'
-          >
-            Submit
-          </Button>
-        </Form>
-      )}
-    </Formik>
+    <div>
+      <Login />
+    </div>
   );
 }
